[ SFT FILE BSP ] moving already existing internal posix FILE IO solution out to a FILE BSP solution. API also added in fiel xi_bsp_io_fs.h. Resource Manager’s file system acces is now channeled to this new FILE BSP. FILE BSP only exists for posix for now.
