+++ src/libxively/io/fs/xi_fs_api.h
@@ -10,47 +10,8 @@
 #include <stdlib.h>
 #include <stdint.h>
 
-#include "xi_err.h"
-
-typedef intptr_t xi_fs_resource_handle_t;
-
-#define XI_FS_INVALID_RESOURCE_HANDLE -1
-#define xi_fs_init_resource_handle() XI_FS_INVALID_RESOURCE_HANDLE
-
-/**
- * @enum xi_fs_resource_type_t
- * @brief describes types of resources that are availible through fs API the
- * types were created in order to differenciate types based on their security
- * class
- */
-typedef enum {
-    XI_FS_CERTIFICATE = 0,
-    XI_FS_CREDENTIALS,
-    XI_FS_CONFIG_DATA
-} xi_fs_resource_type_t;
-
-/**
- * @enum xi_fs_open_flags_t
- */
-typedef enum {
-    XI_FS_OPEN_READ   = 1 << 0,
-    XI_FS_OPEN_WRITE  = 1 << 1,
-    XI_FS_OPEN_APPEND = 1 << 2,
-} xi_fs_open_flags_t;
-
-/*
- * @name xi_fs_stat_s
- * @brief information returned via stat call
- */
-typedef struct xi_fs_stat_s
-{
-    size_t resource_size;
-} xi_fs_stat_t;
-
-#define xi_fs_init_stat()                                                                \
-    {                                                                                    \
-        0                                                                                \
-    }
+#include <xi_err.h>
+#include <xi_bsp_io_fs.h>
 
 /* The size of the buffer to be used for reads */
 extern const size_t xi_fs_buffer_size;
